packages/reference-implementations/python-demo/requirements.txt
langchain==0.3.14
langchain-openai==0.2.14
langgraph>=0.2.75
langchain-core>=0.3.29
python-dotenv==1.0.1
websockets==12.0
requests==2.31.0
google-search-results==2.4.2
aiohttp==3.9.3
pydantic==2.10.5
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0
beautifulsoup4==4.12.3
